requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.7.2
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-phonenumber-field==8.0.0
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
